requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
mysqlclient==2.2.5
pycparser==2.22
PyMySQL==1.1.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Flask==2.2.5
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.3
mysql-connector-python==8.1.0
flasgger==0.9.5
